feat(tenant): implement gRPC services for organization management
- Added gRPC handlers for organization operations including Create, List, Get by ID, Update, Delete, Archive, Restore, and Transfer Ownership.
- Introduced organization member and invitation management with respective handlers for creating, listing, accepting, and canceling invitations.
- Implemented domain models for Organization, OrganizationMember, OrganizationMemberInvitation, and their relationships.
- Added mapping functions for converting domain models to gRPC message types.
- Configured logging and context utilities for enhanced error handling and request tracking.

+++ internal/tenant/models/v1/project.model.go
@@ -21,8 +21,8 @@ type Project struct {
 
 	CreatedBy uuid.UUID `gorm:"type:uuid;not null;index"`
 
-	MaxAPIKeys int `gorm:"column:max_api_keys;default:5;check:max_api_keys > 0 AND max_api_keys <= 10"`
-	MaxMembers int `gorm:"default:5;check:max_members > 0 AND max_members <= 10"`
+	MaxAPIKeys uint8 `gorm:"column:max_api_keys;default:5;check:max_api_keys > 0 AND max_api_keys <= 10"`
+	MaxMembers uint8 `gorm:"default:5;check:max_members > 0 AND max_members <= 10"`
 
 	Members []ProjectMember `gorm:"foreignKey:ProjectID;constraint:OnDelete:CASCADE"`
 	APIKeys []ApiKey        `gorm:"foreignKey:ProjectID;constraint:OnDelete:CASCADE"`
